WIP in case laptop dies

+++ include/peripheral/dma_serial.h
@@ -19,13 +19,13 @@ typedef struct {
 	UART_HandleTypeDef huart;
 	DMA_HandleTypeDef hdma;
 	uint8_t* rx_buffer;
-	uint8_t rx_length;
+	uint16_t rx_length;
 	uint8_t const * rx_tail_ptr;
 } DMA_SerialHandle;
 
-uint8_t dma_serial_count(DMA_SerialHandle* handle);
+uint16_t dma_serial_count(DMA_SerialHandle* handle);
 uint8_t dma_serial_getc(DMA_SerialHandle* handle);
-uint8_t dma_serial_read(DMA_SerialHandle* handle, uint8_t* buffer, uint8_t position, uint8_t length);
+uint8_t dma_serial_read(DMA_SerialHandle* handle, uint8_t* buffer, uint16_t position, uint16_t length);
 
 #ifdef __cplusplus
 }
